Phase 3: adopt uv, add CI, and make mypy clean/blocking
Tooling:
- Switch dependency management to uv: add uv.lock, and drive install,
  lint, format, type, and test through `uv sync` / `uv run` in the
  Makefile.
- Add .github/workflows/ci.yml running lint + type-check + tests on push
  and pull_request (pinned to Python 3.12 via UV_PYTHON).
- Pin the ruff rule set explicitly to E, F, I (keeping the E501/E741
  ignores) so lint is stable across ruff releases; extend-select rode on
  ruff's default set, which changed in 0.16.
- Relax requires-python to >=3.11 (was ==3.12.*), consistent with the
  ruff/mypy 3.11 targets.
- Rename the `delview` optional-dependency group to `gui`; it now also
  covers timeInject, whose PySide6/opencv deps were undeclared.
- Drop the obsolete `ruff[jupyter]` dev dep (no such extra in modern ruff).

Type checking (now clean and blocking in CI):
- DataPoint.hash is Optional and mtime is float, matching how the scan
  pipeline actually populates them; ScanResult fields and AnalysisReport
  tuple types updated to match, with asserts narrowing the never-None
  cases. Behavior is unchanged (guarded by the Phase 2 tests).

+++ src/cais/cli.py
@@ -16,7 +16,7 @@ from photomanager.common.logging import setup_logging
 logger = logging.getLogger(__name__)
 
 
-def format_size(size_bytes: int) -> str:
+def format_size(size_bytes: float) -> str:
     """Converts bytes to a human-readable format."""
     for unit in ["B", "KB", "MB", "GB", "TB"]:
         if size_bytes < 1024.0:
